refactor(drive): Improve OAuth client dependency management
