+++ infrastructure/better-auth/auth-client.ts
@@ -1,6 +1,6 @@
+import type { auth } from "@/infrastructure/better-auth/auth";
 import { inferAdditionalFields } from "better-auth/client/plugins";
 import { createAuthClient } from "better-auth/react";
-import type { auth } from "@/infrastructure/better-auth/auth";
 
 export const authClient = createAuthClient({
   baseURL: "http://localhost:3000",
