+++ infrastructure/better-auth/auth.ts
@@ -2,8 +2,6 @@ import { PrismaClient } from "@/infrastructure/prisma/generated/client";
 import { betterAuth } from "better-auth";
 import { prismaAdapter } from "better-auth/adapters/prisma";
 import { nextCookies } from "better-auth/next-js";
-// import { OAuth2Client } from "google-auth-library";
-// import { google } from "googleapis";
 
 const client = new PrismaClient();
 
@@ -25,17 +23,3 @@ export const auth = betterAuth({
     },
   },
 });
-
-// const oauth2Client = new OAuth2Client(
-//   "VOTRE_CLIENT_ID",
-//   "VOTRE_CLIENT_SECRET",
-//   "http://localhost:3000/callback"
-// );
-
-// const token = await auth.api.getAccessToken({ body: { providerId: "google" } });
-// const refreshToken = await auth.api.oauth2Client.setCredentials({
-//   access_token: token.accessToken,
-//   expiry_date: token.accessTokenExpiresAt,
-// });
-
-// const drive = google.drive({ version: "v3", auth: oauth2Client });
